requirements.simple.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.8.5
aiosignal=1.3.1
async-timeout=4.0.3
attrs=23.1.0
blinker=1.6.2
brotlipy=0.7.0
bzip2=1.0.8
ca-certificates=2023.7.22
cachetools=5.3.1
certifi=2023.7.22
cffi=1.15.1
charset-normalizer=3.2.0
click=8.1.7
cryptography=41.0.3
frozenlist=1.4.0
google-api-core=2.11.1
google-api-python-client=2.97.0
google-auth=2.22.0
google-auth-httplib2=0.1.0
google-auth-oauthlib=1.0.0
googleapis-common-protos=1.60.0
httplib2=0.22.0
idna=3.4
libabseil=20230125.3
libcxx=16.0.6
libexpat=2.5.0
libffi=3.4.2
libprotobuf=4.23.4
libsqlite=3.42.0
libzlib=1.2.13
multidict=6.0.4
ncurses=6.4
oauthlib=3.2.2
openssl=3.1.2
pip=23.2.1
protobuf=4.23.4
pyasn1=0.4.8
pyasn1-modules=0.2.7
pycparser=2.21
pyjwt=2.8.0
pyopenssl=23.2.0
pyparsing=3.1.1
pysocks=1.7.1
python=3.11.4
python_abi=3.11
pyu2f=0.1.5
readline=8.2
requests=2.31.0
requests-oauthlib=1.3.1
rsa=4.9
setuptools=68.1.2
six=1.16.0
tk=8.6.12
typing-extensions=4.7.1
typing_extensions=4.7.1
tzdata=2023c
uritemplate=4.1.1
urllib3=1.26.15
wheel=0.41.1
xz=5.2.6
yarl=1.9.2
